pyproject.toml
[project]
name = "voice-orchestrator"
version = "0.1.0"
description = "Orchestrator for the VOICE project."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0.1,<=8.2.1",
    "loguru>=0.7.3",
    "paramiko<3.4",
    "python-dotenv>=1.2.1",
    "runpod>=1.7.13",
    "sshtunnel>=0.4.0",
    "types-paramiko>=4.0.0.20250822",
    "types-requests>=2.32.4.20250913",
    "zenml[local]>=0.91.0",
]

[project.scripts]
dashboard = "voice_orchestrator.cli.dashboard:main"
run = "voice_orchestrator.cli.run:main"

[tool.setuptools]
package-dir = {"" = "src"}

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "pip>=25.3",
    "pre-commit>=4.3.0",
    "pydoclint>=0.7.6",
    "ruff>=0.14.0",
]

[tool.ruff]
line-length = 90
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "D"]
ignore = ["D203", "D212"]

[tool.pydoclint]
style = "sphinx"
exclude = [
  "src/voice_finetune/config.py",
]
